+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - parameter_openapi_schema passed during prefect deployment registration via rest api
+- resolve variables functionality added
 
 [0.0.6]: https://github.com/yourusername/prefect-deployments-toolkit/releases/tag/v0.0.6